file format to only include the generic dropdown element
