--- frontend/src/main/components/School/SchoolDropdown.js
@@ -1,15 +1,0 @@
-import OurAddDropdownForm from '../OurAddDropdownForm';
-
-function SchoolDropdown({ initialSchool, schools}){
-
-    return (
-        <OurAddDropdownForm
-            content={schools}
-            label="Name"
-            basis={initialSchool}
-            testId="SchoolDropdown-name"
-        />
-    )
-}
-
-export default SchoolDropdown;
